Refactor: Remove unused models and implement new models for assignments, quizzes, blogs, categories, certificates, comments, notifications, payments, progress tracking, and instructor management
- Deleted QuizSubmission, Review, Student, and User models.
- Added new models: AssignmentSubmission, Assignment, Blog, Category, Certificate, Comment, Notification, Payment, Progress, Quiz, and Instructor.
- Implemented pre-save and post-save middleware for various models to handle unique slugs, notifications, and data integrity.
- Enhanced course model to manage relationships with students, instructors, and associated documents.
- Improved error handling and validation across models.
