+++ lib/actions/ai.action.js
@@ -6,9 +6,9 @@ import { GoogleGenAI } from "@google/genai";
 
 import { objectId } from "../utils";
 
-import Category from "@/models/Category";
-import Course from "@/models/Course";
-import Progress from "@/models/Progress";
+import Category from "@/models/category.model";
+import Course from "@/models/course.model";
+import Progress from "@/models/progress.model";
 
 // Using Google Gemini AI for message generation
 const { GEMINI_API_KEY } = process.env;
