+++ lib/actions/assignment.action.js
@@ -6,9 +6,9 @@ import { revalidatePath } from "next/cache";
 import dbConnect from "../db-connect";
 import { objectId } from "../utils";
 
-import Assignment from "@/models/Assignment";
-import AssignmentSubmission from "@/models/AssignmentSubmission";
-import Student from "@/models/Student";
+import AssignmentSubmission from "@/models/assignment-submission.model";
+import Assignment from "@/models/assignment.model";
+import Student from "@/models/student.model";
 
 export async function createAssignment({ courseId, data, path }) {
   try {
