+++ lib/actions/blog.action.js
@@ -6,9 +6,9 @@ import { revalidatePath } from "next/cache";
 import dbConnect from "../db-connect";
 import { objectId } from "../utils";
 
-import Blog from "@/models/Blog";
-import Category from "@/models/Category";
-import Comments from "@/models/Comments";
+import Blog from "@/models/blog.model";
+import Category from "@/models/category.model";
+import Comments from "@/models/comment.model";
 
 export async function createBlog({ blog, path }) {
   try {
