+++ lib/actions/category.action.js
@@ -4,7 +4,7 @@ import { revalidatePath } from "next/cache";
 
 import dbConnect from "../db-connect";
 
-import Category from "@/models/Category";
+import Category from "@/models/category.model";
 
 export async function getCategory(categoryParams = "") {
   try {
