+++ lib/actions/certificate.action.js
@@ -5,7 +5,7 @@ import { auth } from "@clerk/nextjs/server";
 import dbConnect from "../db-connect";
 import { objectId } from "../utils";
 
-import Certificate from "@/models/Certificate";
+import Certificate from "@/models/certificate.model";
 
 export async function getCertificateByStudent({
   page = 1,
