+++ lib/actions/comment.action.js
@@ -4,7 +4,7 @@ import { revalidatePath } from "next/cache";
 
 import dbConnect from "../db-connect";
 
-import Comments from "@/models/Comments";
+import Comments from "@/models/comment.model";
 
 export async function postComment({ blog, user, comment, path }) {
   try {
