+++ lib/actions/course.action.js
@@ -7,13 +7,13 @@ import { revalidatePath } from "next/cache";
 import dbConnect from "../db-connect";
 import { objectId } from "../utils";
 
-import Category from "@/models/Category";
-import Course from "@/models/Course";
-import Instructor from "@/models/Instructor";
-import Lesson from "@/models/Lesson";
-import Module from "@/models/Module";
-import Notification from "@/models/Notification";
-import Student from "@/models/Student";
+import Category from "@/models/category.model";
+import Course from "@/models/course.model";
+import Instructor from "@/models/instructor.model";
+import Lesson from "@/models/lesson.model";
+import Module from "@/models/module.model";
+import Notification from "@/models/notification.model";
+import Student from "@/models/student.model";
 
 // get all Courses for courses page
 export async function getCourses({
