+++ lib/actions/curriculum.action.js
@@ -5,9 +5,9 @@ import { revalidatePath } from "next/cache";
 import dbConnect from "../db-connect";
 import { objectId } from "../utils";
 
-import Course from "@/models/Course";
-import Lesson from "@/models/Lesson";
-import Module from "@/models/Module";
+import Course from "@/models/course.model";
+import Lesson from "@/models/lesson.model";
+import Module from "@/models/module.model";
 
 // Server action for adding course curriculum
 export async function addCourseCurriculum({ courseId, data, path }) {
