+++ lib/actions/instructor.action.js
@@ -6,8 +6,8 @@ import { revalidatePath } from "next/cache";
 import dbConnect from "../db-connect";
 import { objectId } from "../utils";
 
-import Instructor from "@/models/Instructor";
-import User from "@/models/User";
+import Instructor from "@/models/instructor.model";
+import User from "@/models/user.model";
 
 export async function getInstructors({ page = 1, limit = 5 }) {
   try {
