+++ lib/actions/instructor.info.action.js
@@ -6,7 +6,7 @@ import { revalidatePath } from "next/cache";
 import dbConnect from "../db-connect";
 import { objectId } from "../utils";
 
-import InstructorInfo from "@/models/InstructorInfo";
+import InstructorInfo from "@/models/instructor-info.model";
 
 // getInstructorInfo for admin
 export async function getInstructorInfo({
