+++ lib/actions/notification.action.js
@@ -4,7 +4,7 @@ import { auth } from "@clerk/nextjs/server";
 import dbConnect from "../db-connect";
 import { objectId } from "../utils";
 
-import Notification from "@/models/Notification";
+import Notification from "@/models/notification.model";
 
 export async function getReceiveNotification({ page = 1, limit = 3 } = {}) {
   try {
