+++ lib/actions/payment.action.js
@@ -5,7 +5,7 @@ import Stripe from "stripe";
 import dbConnect from "../db-connect";
 import { objectId } from "../utils";
 
-import Payment from "@/models/Payment";
+import Payment from "@/models/payment.model";
 
 const stripe = new Stripe(process.env.STRIPE_SECRET_KEY);
