+++ lib/actions/progress.action.js
@@ -3,7 +3,7 @@ import { auth } from "@clerk/nextjs/server";
 
 import { objectId } from "../utils";
 
-import Progress from "@/models/Progress";
+import Progress from "@/models/progress.model";
 
 export async function updateProgress({ lessonId, courseId, moduleId }) {
   try {
