+++ lib/actions/quiz.action.js
@@ -6,9 +6,9 @@ import { revalidatePath } from "next/cache";
 import dbConnect from "../db-connect";
 import { objectId } from "../utils";
 
-import Quiz from "@/models/Quiz";
-import QuizSubmission from "@/models/QuizSubmission";
-import Student from "@/models/Student";
+import QuizSubmission from "@/models/quiz-submission.model";
+import Quiz from "@/models/quiz.model";
+import Student from "@/models/student.model";
 
 export async function createQuiz({ courseId, data, path }) {
   try {
