+++ lib/actions/review.action.js
@@ -6,7 +6,7 @@ import { revalidatePath } from "next/cache";
 import dbConnect from "../db-connect";
 import { objectId } from "../utils";
 
-import Review from "@/models/Review";
+import Review from "@/models/review.model";
 
 export async function saveReview({ reviewData }) {
   try {
