+++ lib/actions/stats.action.js
@@ -5,11 +5,11 @@ import { auth } from "@clerk/nextjs/server";
 import dbConnect from "../db-connect";
 import { objectId } from "../utils";
 
-import Assignment from "@/models/Assignment";
-import Certificate from "@/models/Certificate";
-import Course from "@/models/Course";
-import Quiz from "@/models/Quiz";
-import Student from "@/models/Student";
+import Assignment from "@/models/assignment.model";
+import Certificate from "@/models/certificate.model";
+import Course from "@/models/course.model";
+import Quiz from "@/models/quiz.model";
+import Student from "@/models/student.model";
 
 // Get last three months course selling data
 export async function courseSellingData() {
