+++ lib/actions/student.action.js
@@ -5,8 +5,8 @@ import { revalidatePath } from "next/cache";
 
 import dbConnect from "../db-connect";
 
-import Student from "@/models/Student";
-import User from "@/models/User";
+import Student from "@/models/student.model";
+import User from "@/models/user.model";
 
 export async function updateStudent({ clerkUserId, user, path }) {
   try {
