+++ lib/actions/user.actions.js
@@ -2,7 +2,7 @@
 
 import dbConnect from "../db-connect";
 
-import User from "@/models/User";
+import User from "@/models/user.model";
 
 export async function createUser(userData) {
   try {
